% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lnc.R
\name{col2rownames}
\alias{col2rownames}
\title{Set a column as row names}
\usage{
col2rownames(df, col)
}
\arguments{
\item{df}{A data frame}

\item{col}{A integer, which column chosen to be as row names}
}
\value{
A data frame
}
\description{
Set a column as rownames and remove that column in the data frame
}
