% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lnc.R
\name{custom_lnc}
\alias{custom_lnc}
\title{Function to join customized lncRNA expression with other genes' expression}
\usage{
custom_lnc(lncRNA, cohort.file)
}
\arguments{
\item{lncRNA}{A character string showing the file which stores a lncRNA expression provided by users}

\item{cohort.file}{A character string, path to the file which stores other genes' expression}
}
\value{
A data frame ready for lnc_gsea
}
\description{
Join lncRNA and other genes' expression, prepare a data frame ready for lnc_gsea
}
