% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lnc.R
\name{filter_mean}
\alias{filter_mean}
\title{Filter out low expression genes}
\usage{
filter_mean(df, col, cutoff)
}
\arguments{
\item{df}{A data frame, rows are genes, columns are samples}

\item{col}{A integer showing column number, df[[col]] is not considered for filtering}

\item{cutoff}{A numeric value as expression threshold}
}
\value{
A data frame with genes' expression larger than cutoff
}
\description{
Filter out low expression genes by mean of FPKM smaller than an cutoff
}
