% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{plot_gsea}
\alias{plot_gsea}
\title{plot enriched pathways
A function to plot enriched pathways, y axis is normalized enrichment
score, x axis is sign(NES)*(-log10(FDR))}
\usage{
plot_gsea(
  gsea.df,
  pathway.list = NULL,
  direction = c("both", "pos", "neg"),
  n = 3
)
}
\arguments{
\item{gsea.df}{A data frame, output from sg_gsea function}

\item{pathway.list}{Pathways user provided to label, a character variable}

\item{direction}{A string, "pos" or "neg" or "both" to label positive or negative or both enriched pathways}

\item{n}{A number, defining how many pos or neg or both pathways to label}
}
\value{
A pdf plot with NES as y axis and sign(NES)*(-log10(FDR)) as x axis
}
\description{
plot enriched pathways
A function to plot enriched pathways, y axis is normalized enrichment
score, x axis is sign(NES)*(-log10(FDR))
}
