% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiCompare.R
\name{plot_multiCompare}
\alias{plot_multiCompare}
\title{plot enriched pathways for one lncRNA in multiple cohorts}
\usage{
plot_multiCompare(gseares)
}
\arguments{
\item{gseares}{A list of data frames which stores interested lncRNA's enriched pathway results for multiple cohorts}
}
\value{
A heatmap plot
}
\description{
Visualize the enriched pathways' results in multi-cohorts for interested lncRNA
}
