% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_gsea.R
\name{pre_gsea}
\alias{pre_gsea}
\title{process expression data for sg_GSEA function}
\usage{
pre_gsea(cohort, t_id)
}
\arguments{
\item{cohort}{A character string which is the short name for cohorts in TCGA}

\item{t_id}{A character string of transcript id, either start with ENST or T}
}
\value{
A data frame of lncRNA's transcript and other genes' expression in interested cohort
}
\description{
Find TCGA samples appear both in RNAseq expression (FPKM) data frame and lncRNAs
data frame, if the sample ids are different, use tcga.meta.info to match.
}
\details{
cohort should be one of studies in TCGA, for example, "PRAD" can be used as
         the function's input to stand for prostate cancer; t_id can be from mitranscriptome
         database or reflnc database.
}
