% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_multiCompare.R
\name{pre_multiCompare}
\alias{pre_multiCompare}
\title{Enriched pathways results of multiple lncRNAs per cohort or multiple cohorts per lncRNA}
\usage{
pre_multiCompare(files, compare = c("lncRNA", "cohort"))
}
\arguments{
\item{files}{A list of data frames which stores enriched pathway results for multi-lncRNA per cohort
or multi-cohort per lncRNA}

\item{compare}{A character string to specify comparison among lncRNAs or among cohorts}
}
\value{
A data frame ready for heatmap by plot_compareCohort
}
\description{
Combine the enriched pathways' results of multiple lncRNAs in one cohort
or one lncRNAs in multiple cohort
}
