% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lnc.R
\name{transdf}
\alias{transdf}
\title{Transpose an expression data frame or matrix}
\usage{
transdf(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A transposed data frame
}
\description{
Transpose an expression data frame or matrix and add rownames as its first column 'patient_id'
}
