% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnc_gsea.R
\name{lnc_gsea}
\alias{lnc_gsea}
\title{lncRNA associated Gene Set Enrichment Analysis}
\usage{
lnc_gsea(
  tid_cohort,
  metric = c("cor", "logFC"),
  cor.method = c("pearson", "spearman"),
  genelist = FALSE,
  geneset = NULL,
  pathway = NULL
)
}
\arguments{
\item{tid_cohort}{A output of data frame from pre_gsea function}

\item{metric}{A string of character, showing which metric is used for rank}

\item{cor.method}{"pearson" or "spearman" method for cor metric}

\item{genelist}{TRUE or FALSE for output ordered gene list or not}

\item{geneset}{A string of character, showing the path where gmt is}

\item{pathway}{A string of character, showing one of pathway in the gmt file}
}
\value{
A file with the GSEA results.If pathway is provided, its enrichment plot
        will be produced too. If genelist is TRUE, also get a file for a ranked gene list
         with ranking metrics, which can be used for GSEA desktop app provided by Broad Institute.
}
\description{
TCGA samples are either stratified by lncRNA's expression and then calculate
log2FC of high vs low expressed groups for each genes, or calculate correlation coefficient
between each gene's expression and the lncRNA's expression. Genes are then ordered by logFC
or correlation coefficent in a descending order. The ranked gene list will be used by `fgsea`
for pre-ranked GSEA.
}
