% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnc_gsea.R
\name{lnc_gsea}
\alias{lnc_gsea}
\title{Single lncRNA Gene Set Enrichment Analysis}
\usage{
lnc_gsea(
  tid_cohort,
  metric = c("cor", "logFC"),
  cor.method = "pearson",
  genelist = FALSE,
  geneset = NULL,
  pathway = NULL
)
}
\arguments{
\item{tid_cohort}{A output of data frame from pre_gsea function}

\item{metric}{A string of character, showing which metric is used for rank}

\item{cor.method}{"pearson" or "spearman" method for cor metric}

\item{genelist}{TRUE or FALSE for output ordered gene list or not}

\item{geneset}{A string of character, showing the path where gmt is}

\item{pathway}{A string of character, showing one of pathway in the gmt file}
}
\value{
A dataframe showing the GSEA results,if pathway is provided, its enrichment plot
        will be produced too.
}
\description{
TCGA samples are either stratified by lncRNA's expression and then calculate
logFC of high vs low express groups, or calculate correlation of genes' expression
with lncRNA's expression, genes are ordered by logFC or correlation coefficent
from largest to smallest. The pre-ranked file is used for pre-ranked GSEA using fgsea
}
